kris-io: rename package to ch.difty.kris
