+++ subprojects/kris-io/src/integrationTest/kotlin/com/gmail/gcolaianni5/jris/usage/JRisIoUsageImportSpec.kt
@@ -2,8 +2,8 @@ package com.gmail.gcolaianni5.jris.usage
 
 import ch.difty.kris.domain.RisRecord
 import ch.difty.kris.domain.RisType
-import com.gmail.gcolaianni5.jris.accept
-import com.gmail.gcolaianni5.jris.process
+import ch.difty.kris.accept
+import ch.difty.kris.process
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.InternalCoroutinesApi
 import org.amshove.kluent.shouldHaveSize
