+++ subprojects/kris-io/src/main/kotlin/ch/difty/kris/ImportExtensions.kt
@@ -1,4 +1,4 @@
-package com.gmail.gcolaianni5.jris
+package ch.difty.kris
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import java.io.File
