+++ subprojects/kris-io/src/main/kotlin/ch/difty/kris/JRisIO.kt
@@ -1,6 +1,5 @@
-package com.gmail.gcolaianni5.jris
+package ch.difty.kris
 
-import ch.difty.kris.JRis
 import ch.difty.kris.domain.RisRecord
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
